채팅 유저 목록, 정산하기 추가

+++ src/main/java/com/example/demo/src/user/ChatService.java
@@ -77,7 +77,13 @@ public class ChatService {
     }
 
     public Object deleteChatRoom(Long chatRoomIdx) throws BaseException {
-        return chatDao.deleteChatRoom(chatRoomIdx);
+        int groupPurchaseMembers = chatDao.getGroupPurchaseMembers(chatRoomIdx);
+
+        if (groupPurchaseMembers == 0) {
+            return chatDao.deleteChatRoom(chatRoomIdx);
+        } else {
+            return "정산이 완료되지 않았습니다.";
+        }
     }
